requirements.txt
# University Management System - Final Requirements
# متطلبات نهائية مُوحدة ومُحدثة لنظام إدارة الجامعة
# This file consolidates all requirements from previous versions

# =============================================================================
# CORE DEPENDENCIES - المتطلبات الأساسية
# =============================================================================

# Essential build tools
setuptools>=69.0.0
wheel>=0.42.0

# Core Django Framework
Django==4.2.16
djangorestframework==3.14.0

# =============================================================================
# AUTHENTICATION & SECURITY - الأمان والمصادقة
# =============================================================================

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Security Enhancements
django-ratelimit==4.1.0
django-csp==3.8
django-permissions-policy==4.18.0
django-cors-headers==4.3.1
python-decouple==3.8

# =============================================================================
# DATABASE & ORM - قاعدة البيانات
# =============================================================================

# PostgreSQL support
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# =============================================================================
# API & DOCUMENTATION - واجهات برمجة التطبيقات والتوثيق
# =============================================================================

# API Documentation
drf-yasg==1.21.7
drf-spectacular==0.27.2

# API Utilities
django-filter==23.5
django-extensions==3.2.3

# =============================================================================
# MEDIA & FILE HANDLING - الملفات والوسائط
# =============================================================================

# Image processing
Pillow==10.4.0
python-magic==0.4.27

# Cloud storage (optional)
django-storages==1.14.2
boto3==1.35.0

# =============================================================================
# REPORTS & DATA EXPORT - التقارير وتصدير البيانات
# =============================================================================

# PDF generation
reportlab==4.2.2

# Excel support
openpyxl==3.1.5
xlsxwriter==3.2.0

# Data analysis
pandas==2.2.2
numpy==1.26.4

# =============================================================================
# CACHING & PERFORMANCE - التخزين المؤقت والأداء
# =============================================================================

# Redis cache
redis==5.0.8
django-redis==5.4.0

# Performance optimization
django-cachalot==2.6.3

# =============================================================================
# BACKGROUND TASKS - المهام الخلفية
# =============================================================================

# Celery for background tasks
celery[redis]==5.3.6
django-celery-beat==2.7.0
django-celery-results==2.5.1

# =============================================================================
# EMAIL & NOTIFICATIONS - البريد الإلكتروني والإشعارات
# =============================================================================

# Email backend
django-anymail==11.0

# =============================================================================
# AI & ANALYTICS - الذكاء الاصطناعي والتحليلات
# =============================================================================

# Machine Learning
scikit-learn==1.5.1

# Data visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.22.0

# =============================================================================
# MONITORING & LOGGING - المراقبة والتسجيل
# =============================================================================

# Error tracking
sentry-sdk[django]==2.14.0

# Health checks
django-health-check==3.18.1

# Structured logging
structlog==24.4.0

# =============================================================================
# PRODUCTION & DEPLOYMENT - الإنتاج والنشر
# =============================================================================

# WSGI server
gunicorn==23.0.0

# Static files serving
whitenoise==6.7.0

# Environment management
django-environ==0.11.2

# =============================================================================
# DEVELOPMENT & TESTING - التطوير والاختبار
# =============================================================================

# Testing framework
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0

# Test data factories
factory-boy==3.3.1

# Code coverage
coverage==7.6.1

# Development tools (only install in development)
django-debug-toolbar==4.4.6

# Code formatting and linting
black==24.8.0
flake8==7.1.1
pre-commit==3.8.0

# =============================================================================
# UTILITIES - الأدوات المساعدة
# =============================================================================

# Date/time utilities
python-dateutil==2.9.0
pytz==2024.1

# HTTP requests
requests==2.32.3

# Arabic text support
arabic-reshaper==3.0.0
python-bidi==0.4.2

# =============================================================================
# OPTIONAL FEATURES - الميزات الاختيارية
# =============================================================================

# Advanced features (uncomment if needed)
# django-taggit==5.0.1                    # Tagging system
# django-mptt==0.16.0                     # Tree structures
# django-simple-history==3.7.0            # Model history
# django-activity-stream==2.0.0           # Activity tracking
# channels==4.1.0                         # WebSocket support
# channels-redis==4.2.0                   # WebSocket with Redis
# daphne==4.1.2                           # ASGI server
# django-haystack==3.3.0                  # Search framework
# elasticsearch==8.15.0                   # Search backend
# django-rosetta==0.10.0                  # Translation interface
# django-modeltranslation==0.18.11        # Model translation

# =============================================================================
# NOTES - ملاحظات
# =============================================================================

# To install development dependencies:
# pip install -r requirements.txt
# pip install django-debug-toolbar  # for development only

# To install production dependencies only:
# pip install -r requirements.txt --only-prod

# For Docker deployment, all dependencies are included
# For cloud deployment, consider using specific versions for your platform